Refactor face recognition to use Bayesian classifier
Replaced SVM-based face recognition and training logic with a Bayesian classifier (MultinomialNB). Updated Predict.py and Train.py to use new model loading, feature normalization, and prediction logic. Removed legacy v2 files, reorganized file structure, and added a README for usage documentation. Updated model and data paths for consistency, and migrated model assets to the bayes trainer directory.
